// frontend/src/generated/types/paymentInfoDetailResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface PaymentInfoDetailResponse {
  sysId?: string;
  orderNum?: number;
  items?: string;
  orderStatus?: string;
  cancelReason?: string;
  cancelDescription?: string;
}